@@ -1,22 +1,22 @@
 name: firestore_ref
 description: Firestore reference/document typed acccess helpers. You can write organized safe code by this.
-version: 0.9.2+4
+version: 0.10.0
 homepage: https://github.com/mono0926/flutter_firestore_ref
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
+  cloud_firestore: ">=0.17.0-0"
+  disposable_provider: ">=2.3.0"
   flutter:
     sdk: flutter
-  cloud_firestore: ">=0.14.0 <1.0.0"
-  disposable_provider: ">=2.2.0 <3.0.0"
-  json_annotation: ">=3.0.0"
-  simple_logger: ">=1.7.0 <2.0.0"
-  meta: ">=1.1.8 <2.0.0"
-  quiver: ">=2.1.3 <3.0.0"
-  rxdart: ">=0.24.1 <1.0.0"
-  subscription_holder: ">=2.0.0 <3.0.0"
+  json_annotation: ">=4.0.0"
+  meta: ">=1.3.0"
+  quiver: ">=3.0.0-0"
+  rxdart: ">=0.26.0-0"
+  simple_logger: ">=1.8.0"
+  subscription_holder: ">=2.1.0"
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic_mono: any
+  pedantic_mono:
 flutter:
