@@ -1,6 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firestore_ref/firestore_ref.dart';
-import 'package:meta/meta.dart';
 
 class TimestampField {
   static const createdAt = 'createdAt';
@@ -8,7 +7,7 @@ class TimestampField {
 }
 
 Map<String, dynamic> replacingTimestamp({
-  @required Map<String, dynamic> json,
+  required Map<String, dynamic> json,
 }) =>
     <String, dynamic>{
       ...json,
@@ -17,9 +16,9 @@ Map<String, dynamic> replacingTimestamp({
       TimestampField.updatedAt: FieldValue.serverTimestamp(),
     };
 
-DateTime parseTimestamp({
-  @required Map<String, dynamic> json,
-  @required String key,
+DateTime? parseTimestamp({
+  required Map<String, dynamic> json,
+  required String key,
 }) {
-  return (json[key] as Timestamp)?.toDate();
+  return (json[key] as Timestamp?)?.toDate();
 }
